+++ TikTokTrainer/Camera/CameraPreview.swift
@@ -1,0 +1,30 @@
+//
+//  CameraPreview.swift
+//  TikTokTrainer
+//
+//  Created by David Sadowsky on 1/29/21.
+//
+
+import Foundation
+import AVFoundation
+import SwiftUI
+
+struct CameraPreview: UIViewRepresentable {
+    @ObservedObject var camera: CameraModel
+
+    func makeUIView(context: Context) -> some UIView {
+        let view = UIView(frame: UIScreen.main.bounds)
+
+        camera.preview = AVCaptureVideoPreviewLayer(session: camera.session)
+        camera.preview.frame = view.frame
+
+        camera.preview.videoGravity = .resizeAspectFill
+        view.layer.addSublayer(camera.preview)
+
+        return view
+    }
+
+    func updateUIView(_ uiView: UIViewType, context: Context) {
+
+    }
+}
